EDC20 Implemented Average Rating

+++ book-rating-service/src/main/java/com/edc/pps/rating/dto/RatingResponse.java
@@ -9,7 +9,7 @@ public class RatingResponse {
     private Long ratingId;
     private Long bookId;
     private Long userId;
-    private Integer ratingValue;
+    private Double ratingValue;
 
     @Override
     public boolean equals(Object o) {
